HHH-18868 Renamed test classes
