+++ hibernate-core/src/test/java/org/hibernate/orm/test/idclass/PKey.java
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.orm.test.metamodel.hhh18868;
+package org.hibernate.orm.test.idclass;
 
 import java.io.Serializable;
 import java.util.Objects;
